Fix logic for fully_resolve
Previously fully_resolve was on stac_object, which operated the same
for catalogs and items. Catalogs were not calling fully_resolve on
links that were previously resolved themselves, causing the recursive
resolution strategy to fail if links were previously resovled. Fixed
this exposed an issue where Item's full_resolve should not be
recursive - or else a situation occurs where trying to fully resolve
the item's collection from the item causes an infinite loop. This
fixes the issue by splitting the implementation of fully_resolve to
handle Catalogs and Items separately.