Add tests for version extension that test full_copy.
This commit adds two tests, one for each of the Item and Collection
extensions of the Version extension. The test checks that when items
or collections are linked by the version links (predecessor, successor,
or latest_version), that a full copy will end up with a catalog that
has those links point to the correct object instances. If this test
were to fail, that would indicate that the STAC object linked to by
the version links was a separate copy of item or collection, which is incorrect.

Redo of https://github.com/stac-utils/pystac/pull/193/commits/defd7efb1180b79bcef0f5b67b2e3fd816fd5bba
by lossyrob