Fix validation issue with tuples vs lists.
Shapely can produce GeoJSON mappings that use tuples in the coordinate
sequences of geometries instead of lists. A direct validation of the
dict for GeoJSON will produce errors, since coordinates are expected
to be list-based. This commit adds a translation to and from JSON
serialization in order to work around this and other issue which may
appear if a JSON serialization would pass validation but the STAC dict
may not.