Ensure resolve object cache is set before set_self_href is called
This commit modifies the order of the logic in the Catalog.__init__
method to ensure the resolved object cache is established before a
set_self_href call is made, which occurs if a non-None href is
supplied. The logic of the Collection's set_self_href method requires
the resolution cache be created on the object, which this change
ensures.

Fixes #182