Allow for unnormalized path prefixes like /vsitar/
The logic in make_absolute_href uses os.path.abspath, which normalizes
paths. This lets us translate joined paths like
/home/user/../catalog.json into proper paths. However, normalizing the
path will also modify the root path being joined to in ways that may
not be desirable. Specifically in the case of a /vsitar/ prefix, the
double slashes that are required for proper GDAL reading are removed
by normalization. This commit modifies the logic in
make_absolute_href to detect when the original start path is modified
by a call to abspath, and if so replaces the modified version with the
original version.

Fixes #180