Refactor caching to utilize HREFs when available.
This commit modifies the caching strategy of PySTAC away from being
based simply on the ID of the STACObject. This can cause issues as
STAC objects do not necessarily have unique IDs, though best practices
say unique IDs are prefered.

This change modifies the cache to be based off of the self HREF first,
and if none is present, based off of not just the object's ID but the
ID of all of it's parents.

The caching is updated any time set_self_href is called on a STAC
object so the cache is kept up-to-date when new HREFs are set on
objects.

Fixes #92