Allow for different layouts in normalize_hrefs
This commit refactors normalize_hrefs to be based on a layout
strategy. This does not change the default behavior.

normalize_hrefs is removed from Item as the method is intended for use
on a Catalog.

The "fully_resolve" method was causing performance issues and is no
longer needed based on this refactor, and so is dropped. A
"resolve_links" method takes its place, which only resolves the stac
object links of the target object (and does not recurse down to
connected objects).