Fix issue in merging common props when no stac_extensions present
The logic around determining when to attempt to merge collection
properties into items was broken. This should only happen for STAC
0.9.0 if the Item implements the 'commons' extension. The logic was
checking for that extension in an existing list, but if the JSON did
not have a 'stac_extensions' property it was performing the
merge. This changes it to avoid the common property merge logic if no
stac_extensions property exists.