Make error more clear when CollectionCache caches non-Collection
The logic in the common properties merging functionality expects a
collection (or a dict representing collection JSON) be cached
according to the ID associated with the collection property in an
Item. If something else is cached there (for example, if a collection
was read in as a Catalog) this produced a confusing error. This
changes the logic to throw a ValueError in that case.