Ensure proper caching in Catalog and Collection from_dict
This commit changes the behavior of the from_dict methods of Catalog
and Collection to pass in the href to the constructor so that caching
logic utilizes the HREF correctly.

This change will change behavior in that a catalog or collection that
is read in with a HREF that is different then the stated self HREF on
the object would previously use the HREF in the link. The to_dict
methods will now set the given HREF as the self link if provided, even
if there's an existing, differing self link. This should only affect
cases when there are incorrect self links in STAC objects.