Add catalog_type to Catalog to track catalog type of read STACs.
Before this change, the save method on Catalog always requires a
CatalogType supplied as an argument. This change determines the
catalog type of a catalog that is read in by examining its Links and
carries that information on the object, so that if a catalog_type is
set on the catalog, it does not need to be supplied as an argument for
the save call.