Modify Item set_self_href to ensure asset HREFs do not break.
Previous to this change, using set_self_href on Item could cause
assets with relative HREFs to lose a pointer to the actual asset
files. This was not true in the case of using nomralize_hrefs on a
catalog containing the item, but calling set_self_href directly on the
item could cause this problem to occur. This change moves the logic to
ensure relative asset HREFs do not break when changing the Item's HREF
into an overridden version of set_self_href on Item.