Handle setting new versions correctly in version extension.
Prior to this change, setting the 'latest', 'successor', or
'predecessor' would add links, but not clear any old links if there
were already versions set. This ensures that the links are cleared
before setting any new version of the object.

This change also ensures that setting a version to 'None' will clear the links,
allowing users to remove a version from an item or collection.

Fixes #210