Modify asset HREFs based on catalog type.
Previously only link href types (absolute vs relative) were modified
as part of the `Catalog.save` process based on the catalog types: for
ABSOLUTE_PUBLISHED, all link hrefs were made absolute, etc. The asset
HREFs were not modified, which was counter to the expected result
outlined in https://github.com/stac-utils/stactools/issues/31.

This change causes asset HREFs to also be changed to absolute or
relative based on the catalog type. Note that if the asset HREF cannot
be made relative, e.g. the asset HREF points to an HTTP location and
the item is being stored locally, then the HREF is left unchanged.