Fix issue with proj extension setters handling None values.
Before this change, setting a projection extension value to None would
break if the value wasn't required; the proper way to handle this case
is to pop the property key or not set it. This fixes an issue brought
up in #268 around using apply with the only required property (epsg).