New extension architecture in place, refactor eo, file and label.
Introdues "ExtensionHooks" which are not necessary for an extension
but are registered if the extension needs to modify the behavor of
PySTAC in it's processes. The two places extensions can currently hook
into are during object resolution, where they can add link reltypes
that indicate STACObject links based on the extension; also migration
- the logic for migrating extensions will be moved from the migrate
subpackage to each extension itself.