Remove type parameter from summary methods.
In #337 we are removing type parameters from methods as it conflicts
with mypy checking and doesn't add a lot of value. This change is
implemented here to work around some type checker issues that could
have been solved a different way, but removing the type parameter as
that would have been future work anyway.