Use Protocol from typing_extensions for pre-3.8 Python
This commit adds an extra dependency to PySTAC when installed in
environments with versions of Python before 3.8. The 'Protocol' class
in `typing` was made available in the core module as of 3.8, but
previous versions can use it from the `typing_extensions`
package. setup.py is changed to install typing_extensions for pre-3.8
versions, and a conditional import is used to import the Protocol
class from the correct package.