Allow root to be set during to_dict
There was existing unused arguments for seting the root on
to_dict. This commit implements and tests that the root is set to the
parameter if present, and adds the parameter to ItemCollection, which
itself does not have a root but instead passes the root onto each of
the Items parsed during the to_dict call.

Related to #546