Allow resolved self links
When getting a STAC object's self href, there was a cast that assumed
that the object's self link's target was a string, which was not always
the case (the self link's target could be a string or another STAC
object). This commit removes that cast by splitting a link's target into
two separate "private" attributes, `_target_href` and `_target_object`.

Fixes #499.