Enable structural typing of RangeSummary to allow for datetime
Structural typing requires that the parameter names match. For numeric
types like float, the __lt__ method takes an "x" parameter. For
datetime, the __lt__ method takes an "other" parameter, which caused
Pyright to show datetime as not matching the bounds of the TypeVar T
when used with RangeSummary.

This modifies the bounds to be a union of Protocol types that allow
either the "x" or "other" parameter