Pin rustc version to nightly-2021-04-27 in CI
When orjson doesn't have a wheel available for a given OS+Python, we
have to build it from source during CI. This requires the rust compiler
(and associated tooling). There was a minor regression in the output of
the `--version` command in one rust tool (`cargo`) which caused the
`orjson` build chain to break. When
https://github.com/rust-lang/cargo/pull/9727 is incorporated back into
nightly Rust releases we can unpin this toolchain and revert it back to
`nightly`.

The check that was failing because of the regression:
https://github.com/PyO3/maturin/blob/5f134e3a4adbbf21f04c10f1dc9722742156f959/maturin/__init__.py#L114-L122