fix: create valid dict without root href
Catalogs and Collections always have a root href. If the root Catalog or
Collection (e.g. itself) does not have a self href, the dictionary serialization
included a root link with href set to `null`, which is invalid. This fix removes
the root link if it doesn't have a valid href.

Not required for items because they don't have a root link by default.

I discovered this problem when trying to write-then-read a very simple
Collection using `save_object`, but the problem proved to be in the dict
serialization.