Add projection v1.2.0 to the list schema URIs in projection.py
The migration of projection to latest version systematically pops the field epsg. If extension URI is pointing to versions 1.0.0 or 1.1.0, it is also changed to point to version 2.0.0. 
However, if schema URI is v1.2.0 the extension schema URI is not changed. I propose to change it as well in that case so that the change is explicit for all versions < 2.0.0. That would help finding out that a change occurred.